Move (un)install code into `soundcraft_installtool` script
Moves code from the `soundcraft.dbus` module which is not
related to running the D-Bus client or the D-Bus service to
a new module `soundcraft.installtool`.  That module then
implements the new script `soundcraft_installtool`.

This split in turn allows us to remove the installation code
(script, module, and data files) after use if we do not need
it any more, e.g. when preparing a binary package for a Linux
distribution.
